linkedin/company_data/requirements.txt
numpy
pandas
selenium
bs4
lxml
azure-storage-blob
undetected_chromedriver 